+++ Runtime/Data Classes/GeneralMusic.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+using System;
+
+namespace AltifoxStudio.AltifoxAudioManager
+{
+    public enum MusicDivision
+    {
+        instant,
+        beat,
+        measure
+    }
+}