+++ Runtime/Static Classes/AltifoxGizmos.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 #if UNITY_EDITOR
 using UnityEditor;
-#endif
+
 
 namespace AltifoxStudio.AltifoxAudioManager
 {
@@ -127,3 +127,4 @@ namespace AltifoxStudio.AltifoxAudioManager
         }
     }
 }
+#endif