+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Fixed a bug where the first loop would end too early
 
 ## [1.0.2] - 30/07/2025
+- You can now create soundscapes, they allow for the random spawning of one shot sound sources within a box